♻️ Trekk funksjon for å validere ingen endringer før revurder fra ut i en felles fil for læremidler og tilsyn barn. Ta i bruk felles interface `PeriodeMedId` som kan brukes av vedtaksperioder tilsyn barn og vedtaksperioder læremidler
